+++ coda-src/repair/repair.cc
@@ -426,7 +426,7 @@ void rep_ReplaceInc(int largc, char **largv)
 	fprintf(stderr, "File %s is not a regular file (and hence cannot be used for repair)\n", mergefile);
 	return;
     }
-    if (!repair_getfid(mergefile, &fixfid, fixrealm, &fixvv, msgbuf, sizeof(msgbuf)) && (fixvv.StoreId.Host != (unsigned long)-1))
+    if (!repair_getfid(mergefile, &fixfid, fixrealm, &fixvv, msgbuf, sizeof(msgbuf)) && (fixvv.StoreId.Host != (unsigned)-1))
 	sprintf(fixpath, "@%x.%x.%x@%s", fixfid.Volume, fixfid.Vnode, fixfid.Unique, fixrealm);
     else strcpy(fixpath, mergefile);
